# Add hashing functions to speed up normal usage