--- a/pyproject.toml
+++ b/pyproject.toml
@@ -2,3 +2,9 @@
 # By default `tox-pip-sync` disables listing the contents of the env for speed.
 # You can turn off this behavior if you want:
 skip_listing = true
+
+# Once we succesfully install the dependencies for a virtual env, we will store
+# a hash of them. If nothing changes we won't attempt to resync the virtual
+# environment, which can save several seconds. You disable this so the
+# environment is synced every time by setting this to `false`
+enable_hashing = true