For: https://github.com/hypothesis/tox-pip-sync/issues/14

This PR adds hashing as a matter of course. This means we will:

 * Create a hash of the dependencies the last time we synced
 * Calculate the hash now
 * Only trigger re-syncing if the two are different

This _**does not detect changes in the tox env itself**_. So if you mess with the tox env manually, it'll stay as it is until the next sync is triggered. You can remove the tox env entirely to "fix" this. I think this is actually handy behavior as you can install things in to test them etc.

You can control this behavior by setting the `enable_hashing` setting in `pyproject.toml`

## Example benefits

To test this I ran `time make format sure` in `h-matchers`. This was run a couple of times to ensure the tox envs were all created first:

 * Time with `tox-pip-sync` from master: 22.3s
 * This branch (first run): 22.2s
 * This branch (second run): 15.6s (30% saving)
 
A second example of a frequently run command without any Make overhead `tox -e tests`:

 * Time with `tox-pip-sync` from master: 5.9s
 * This branch (first run): 5.8s
 * This branch (second run): 3.3s  (43% saving)

## What we hash

Things the hash depends on (and should notice changing):

 * Project files: `pyproject.toml`, `setup.py`, `setup.cfg`
 * Any dependencies specified directly in `tox.ini`
 * Any dependencies in `requirements.txt` files

## Testing notes

Pick a library and an app and repeat for both:

 * `tox -e tests --run-command echo` - Just to get the environment up and running first
 * `.tox/.tox/bin/pip install -e ~/my_projects_dir/tox_pip_sync`
 * `time tox -v -e tests --run-command echo` 
 * You should see a comment about the project being re-synced
 * `time tox -v -e tests --run-command echo` 
 * You should not see a comment about the project being re-synced and if should be faster
 * Try changing any of the things and re-run
 * `time tox -v -e tests --run-command echo` 
 * You should see a comment about the project being re-synced